Free for any use, but you must disclose source code of improvements
This software is open source and depends on contributions from the community. Under the GNU GPL v3 License, you are given permission to use this freely for any purpose you like, with no liability and no warranty, as long as you make the source code of any improvements available for all.